修復自動輪替「依隊伍分工指定目標」的負數參數陣列越界：改用無號比較
StrategyTarget.PartyByAssignment 分支原本只檢查 param < spa.Length，缺下界。
param 就是預設／計畫檔裡的 TargetParam，Strategy.cs 反序列化時直接 jtp.GetInt32()，
任何 int 都進得來，所以一個負值（手改或壞掉的計畫檔）會讓 spa[param] 擲
IndexOutOfRangeException —— 而且是在每幀都會走到的 ResolveTargetOverride 上擲。

改成 (uint)param < (uint)spa.Length：負值轉無號會繞成極大值，被同一個比較一併擋掉，
落回這個 switch 本來就存在的「回 null」語意，沒有新增任何行為。

同檔另外兩個由 param 導出的索引已逐一確認不同型、不需要改：
- WorldState.Waymarks[(Waymark)param]：WaymarkState 取值先過 BitMask，而
  BitMask.MaskForBit 自己有 (uint)index < 64u 閘門，且第 8~63 位永遠沒被設過，
  越界只會得到 null 不會擲例外。
- (StrategyPartyFiltering)param 與 (StrategyEnemySelection)param 是列舉轉型，沒有陣列索引。
內層的 WorldState.Party[...] 走的是 PartyState 自己的邊界檢查取值器（越界回 null）。

+++ BossMod/Autorotation/RotationModuleManager.cs
@@ -137,7 +137,10 @@ public sealed class RotationModuleManager : IDisposable
     public Actor? ResolveTargetOverride(StrategyTarget strategy, int param) => strategy switch
     {
         StrategyTarget.Self => Player,
-        StrategyTarget.PartyByAssignment => _prc.SlotsPerAssignment(WorldState.Party) is var spa && param < spa.Length ? WorldState.Party[spa[param]] : null,
+        // 🔴 param 來自預設／計畫 JSON 的 TargetParam（Strategy.cs 直接 GetInt32()），負值進得來。
+        //    舊的 `param < spa.Length` 對 -1 是成立的 ⇒ spa[-1] 擲 IndexOutOfRangeException。
+        //    轉 uint 比較讓負值一起被擋掉（會繞回極大值），落回本來就存在的「回 null」語意。
+        StrategyTarget.PartyByAssignment => _prc.SlotsPerAssignment(WorldState.Party) is var spa && (uint)param < (uint)spa.Length ? WorldState.Party[spa[param]] : null,
         StrategyTarget.PartyWithLowestHP => FilteredPartyMembers((StrategyPartyFiltering)param).MinBy(a => a.HPMP.CurHP),
         StrategyTarget.EnemyWithHighestPriority => Hints.PriorityTargets.MaxBy(RateEnemy((StrategyEnemySelection)param))?.Actor,
         StrategyTarget.EnemyByOID => Player != null && (uint)param is var oid && oid != 0 ? Hints.PotentialTargets.Where(e => e.Actor.OID == oid).MinBy(e => (e.Actor.Position - Player.Position).LengthSq())?.Actor : null,
